+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_99V9_Size.java
@@ -119,7 +119,7 @@ public class PaternDataTypeLevel_99V9_Size extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -138,7 +138,7 @@ public class PaternDataTypeLevel_99V9_Size extends AbstractPattern{
 		//	*	12 Mandatory Sayi
 		else if(abstractTokenInPattern.getSourceFieldName().equals("decimalLength")){
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setDecimalLength( db.intValue());
+			matchedCommandAdd.setDecimalLength( db.longValue());
 			matchedCommandAdd.getParameters().put("decimalLength", matchedCommandAdd.getDecimalLength());
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("REDEFINES")){
@@ -147,7 +147,7 @@ public class PaternDataTypeLevel_99V9_Size extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
