+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_9_SizeV99.java
@@ -120,7 +120,7 @@ public class PaternDataTypeLevel_9_SizeV99 extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -133,7 +133,7 @@ public class PaternDataTypeLevel_9_SizeV99 extends AbstractPattern{
 		//*	12 Mandatory Sayi
 		else if(abstractTokenInPattern.getSourceFieldName().equals("length")){ 
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setLength( db.intValue());
+			matchedCommandAdd.setLength( db.longValue());
 			matchedCommandAdd.getParameters().put("length", matchedCommandAdd.getLength());
 		}
 		
@@ -149,7 +149,7 @@ public class PaternDataTypeLevel_9_SizeV99 extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
