+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_9_Size_V9_Size.java
@@ -135,7 +135,7 @@ public class PaternDataTypeLevel_9_Size_V9_Size extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -147,13 +147,13 @@ public class PaternDataTypeLevel_9_Size_V9_Size extends AbstractPattern{
 		}
 		//12 Mandatory Sayi
 		else if(abstractTokenInPattern.getSourceFieldName().equals("length")){ 
-			matchedCommandAdd.setLength(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLength(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("length", matchedCommandAdd.getLength());
 		}
 		//12 Mandatory Sayi
 		else if(abstractTokenInPattern.getSourceFieldName().equals("decimalLength")){ 
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setDecimalLength( db.intValue());
+			matchedCommandAdd.setDecimalLength( db.longValue());
 			matchedCommandAdd.getParameters().put("decimalLength", matchedCommandAdd.getDecimalLength());
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("REDEFINES")){
@@ -162,7 +162,7 @@ public class PaternDataTypeLevel_9_Size_V9_Size extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
