+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_AA.java
@@ -94,7 +94,7 @@ public class PaternDataTypeLevel_AA extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("A");
 			matchedCommandAdd.getParameters().put("type","String");
@@ -115,7 +115,7 @@ public class PaternDataTypeLevel_AA extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
