+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_S9_Size.java
@@ -110,7 +110,7 @@ public class PaternDataTypeLevel_S9_Size extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -123,7 +123,7 @@ public class PaternDataTypeLevel_S9_Size extends AbstractPattern{
 		//length
 		else if(abstractTokenInPattern.getSourceFieldName().equals("length")){ 
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setDecimalLength( db.intValue());
+			matchedCommandAdd.setDecimalLength( db.longValue());
 			matchedCommandAdd.getParameters().put("length", matchedCommandAdd.getLength());
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("REDEFINES")){
@@ -132,7 +132,7 @@ public class PaternDataTypeLevel_S9_Size extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
