+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_S9_Size_V99.java
@@ -119,7 +119,7 @@ public class PaternDataTypeLevel_S9_Size_V99 extends AbstractPattern{
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -132,7 +132,7 @@ public class PaternDataTypeLevel_S9_Size_V99 extends AbstractPattern{
 		//length
 		else if(abstractTokenInPattern.getSourceFieldName().equals("length")){ 
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setLength( db.intValue());
+			matchedCommandAdd.setLength( db.longValue());
 			matchedCommandAdd.getParameters().put("length", matchedCommandAdd.getLength());
 		}
 		//V999 Mandatory Kelime (V ile basla)
@@ -146,7 +146,7 @@ public class PaternDataTypeLevel_S9_Size_V99 extends AbstractPattern{
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
