+++ CobolPorterParser/src/main/java/tr/com/vbt/cobol/parser/datalayout/patern/PaternDataTypeLevel_S9_Size_V9_Size.java
@@ -136,7 +136,7 @@ ElementDataTypeCobol matchedCommandAdd=(ElementDataTypeCobol) matchedCommand;
 		super.setSatirNumarasi(matchedCommand,currentTokenForMatch, abstractTokenInPattern);if(abstractTokenInPattern.getSourceFieldName()==null){
 		
 		}else if(abstractTokenInPattern.getSourceFieldName().equals("levelNumber")){
-			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).intValue());
+			matchedCommandAdd.setLevelNumber(((Double)currentTokenForMatch.getDeger()).longValue());
 			matchedCommandAdd.getParameters().put("levelNumber", matchedCommandAdd.getLevelNumber());
 			matchedCommandAdd.setDataType("9");
 			matchedCommandAdd.getParameters().put("type","number");
@@ -149,13 +149,13 @@ ElementDataTypeCobol matchedCommandAdd=(ElementDataTypeCobol) matchedCommand;
 		//length
 		else if(abstractTokenInPattern.getSourceFieldName().equals("length")){ 
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setLength( db.intValue());
+			matchedCommandAdd.setLength( db.longValue());
 			matchedCommandAdd.getParameters().put("length", matchedCommandAdd.getLength());
 		}
 		//V999 Mandatory Kelime (V ile basla)
 		else if(abstractTokenInPattern.getSourceFieldName().equals("decimalLength")){  
 			Double db=(Double) currentTokenForMatch.getDeger();
-			matchedCommandAdd.setDecimalLength( db.intValue());
+			matchedCommandAdd.setDecimalLength( db.longValue());
 			matchedCommandAdd.getParameters().put("decimalLength", matchedCommandAdd.getDecimalLength());
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("REDEFINES")){
@@ -164,7 +164,7 @@ ElementDataTypeCobol matchedCommandAdd=(ElementDataTypeCobol) matchedCommand;
 		}
 		else if(abstractTokenInPattern.getSourceFieldName().equals("VALUE")){
 			System.out.println(currentTokenForMatch.toString());
-			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Integer) {
+			if(((KeyValueOzelKelimeToken)currentTokenForMatch).getValue() instanceof Long) {
 				Long valueAsInt= (Long) ((KeyValueOzelKelimeToken)currentTokenForMatch).getValue();
 				matchedCommandAdd.setValueAsInt(valueAsInt);
